% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing.r
\name{parse_gps}
\alias{parse_gps}
\title{Import Running Data}
\usage{
parse_gps(path)
}
\arguments{
\item{path}{Relative path to single .gpx file in subfolder or export directory.}
}
\value{
Tibble with gps-position data (lat, lon, elevation), time_span,

(cumulative) distance (time_span, duration, dist, dist_cumsum).
}
\description{
Parse single .gpx file from runtastic export zipfile.
}
\examples{
parse_gps(path = 'cad66c60-474d-4c2c-baa0-699d784de564.gpx')
}
